app/requirements.txt
flask
python-dotenv
newrelic
newrelic-telemetry-sdk
